requirements.txt

requests
firebase-admin

python-dotenv